+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. Dates are d
 > 2022-04-21
 
 - Add `throttle` option
+- Pass `contextData` through to `before` function call
 
 ### [1.8.1](https://github.com/doesdev/get-scrud/compare/1.8.0...1.8.1)